+++ src/auto_placement.h
@@ -2,6 +2,9 @@
 #ifndef __AUTOAPP_H__
 #define __AUTOAPP_H__
 
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
 #include <windows.h>
 
 #include <wx/arrstr.h>