+++ src/dialog_hotkeys.cpp
@@ -1,3 +1,6 @@
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
 #include <windows.h>
 #include <wx/statline.h>
 #include <wx/wx.h>