+++ src/frame_hook.h
@@ -2,6 +2,9 @@
 #define __MAIN_FRAME_H__
 
 #include <vector>
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
 #include <windows.h>
 #include <wx/wx.h>