+++ src/hook.h
@@ -1,6 +1,9 @@
 #ifndef __HOOK_HEADER_H__
 #define __HOOK_HEADER_H__
 
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
 #include <windows.h>
 
 #ifdef BUILD_DLL