+++ src/layout_manager.h
@@ -7,6 +7,9 @@
 
 #include <vector>
 
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
 #include <windows.h>
 
 #include "settingsmanager.h"