+++ src/list_windows.h
@@ -1,6 +1,9 @@
 #ifndef __LISTER_FENETRE_H__
 #define __LISTER_FENETRE_H__
 
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
 #include <windows.h>
 
 class ListWindows {