+++ src/update_thread.cpp
@@ -1,5 +1,8 @@
 
 #include <time.h>
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
 #include <windows.h>
 #include <wx/protocol/http.h>
 #include <wx/sstream.h>